리팩토링: worker 슬롯 관리 공통화
